move reactive extension to Manager class
