+++ lib/src/sprinkle_extension.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:rxdart/rxdart.dart';
 
 import 'provider.dart';
 
@@ -24,31 +23,6 @@ extension SprinkleExtension<T> on BuildContext {
   bool get isLandscape => orientation == Orientation.landscape;
 
   bool get isPortrait => orientation == Orientation.portrait;
-}
-
-extension SprinkleInt on int {
-  BehaviorSubject<int> get reactive => BehaviorSubject<int>.seeded(this);
-}
-
-extension SprinkleString on String {
-  BehaviorSubject<String> get reactive => BehaviorSubject<String>.seeded(this);
-}
-
-extension SprinkleDouble on double {
-  BehaviorSubject<double> get reactive => BehaviorSubject<double>.seeded(this);
-}
-
-extension SprinkleBool on bool {
-  BehaviorSubject<bool> get reactive => BehaviorSubject<bool>.seeded(this);
-}
-
-extension SprinkleList<T> on List<T> {
-  BehaviorSubject<List<T>> get reactive =>
-      BehaviorSubject<List<T>>.seeded(this);
-}
-
-extension SprinkleObject<T> on T {
-  BehaviorSubject<T> get reactive => BehaviorSubject<T>.seeded(this);
 }
 
 extension SprinkleWidget on Widget {
